Use highwater OFPort allocation (dovesnap will ask OVS to assign new interfaces, the lowest possible free OFPort number). This minimizes the number of preallocated FAUCET ports.
We ask OVS for allocated port numbers using ovs-ofctl. While it is possible to obtain this information via ovsdb,
it is much more complex - we have to ask the Bridge table for ports, the Ports table for Interfaces, and finally
the Interfaces table for OFPort numbers. Any of those transactions might fail or not be atomic.

ovs-ofctl uses a Unix domain socket for OpenFlow communication, so there is no extra OpenFlow TCP port necessary.
It would be good to remove the rest of dovesnap's legacy use of the ovsdb library and have it run ovs-vsctl over
a domain socket as well, which would reduce OVS' network attack surface further by not having to listen on the management port.