Make Join() and CreateEndpoint() non blocking again, but retain the head start on OFPort provisioning in CreateEndpoint.
We cannot block a Join() or CreateEndpoint(), ever, because Docker holds locks on the container and network associated with the new endpoint until either call has returned.

In the case of a container connected to multiple Dovesnap networks, the inspect call on the container for the first network, will block because there is a second CreateEndpoint() and Join() pending.

We cannot block the first Join() to avoid the second pending Join() either, because the inspect will still block because the first Join() hasn't returned.

We have no choice but to allow Docker to proceed and complete our provisioning in the background, but at least we can speed up this work with preallocate_ports and start provisioning OVS from the first CreateEndpoint() call.